+++ src/test/java/ru/netology/springmvc/service/FileServiceTest.java
@@ -40,7 +40,7 @@ class FileServiceTest {
 
     private final Files fileEntity = Files.builder()
             .filecontent(file.getBytes())
-            .size(file.getSize())
+            .size((int)file.getSize())
             .filename(file.getName())
             .type(file.getContentType())
             .userid(userId1).build();
